% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_Omega}
\alias{get_Omega}
\title{Covariance matrix estimation}
\usage{
get_Omega(u0, S_X1, S_Y1, S_Y1X1)
}
\description{
Covariance matrix from the asymptotic distribution derived in Gossmann et. al. (2018)
}
\details{
Get the covariance matrix defined in Theorem 1 of Gossmann et. al. (2018)
"FDR-Corrected Sparse Canonical Correlation Analysis with Applications to Imaging Genomics"
(https://arxiv.org/pdf/1705.04312).
}
