% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA.R
\name{L1_CCA_CV}
\alias{L1_CCA_CV}
\title{L1-penalized SCCA with CV}
\usage{
L1_CCA_CV(X, Y, lambda_X, lambda_Y, n_folds = 5)
}
\description{
L1-penalized SCCA with hyperparameter choice via k-fold cross-validation
}
\details{
The considered search space is effectively one-dimensional by considering
penalty pairs (lambda_X[1], lambda_Y[1]), (lambda_X[2], lambda_Y[2]), etc.
The number of cross-validation folds can be specified.
}
