% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA.R
\name{L1_CCA_loocv}
\alias{L1_CCA_loocv}
\title{L1-penalized SCCA with LOOCV}
\usage{
L1_CCA_loocv(X, Y, lambda_X, lambda_Y, num_cores)
}
\description{
L1-penalized SCCA with hyperparameter choice via leave-one-out cross-validation
}
