% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCA.R
\name{L1_CCA_with_sparsity_bound}
\alias{L1_CCA_with_sparsity_bound}
\title{L1-penalized SCCA with sparsity bounds}
\usage{
L1_CCA_with_sparsity_bound(X, Y, bound_X, bound_Y, tolerance, verbose = FALSE,
  max_iter = 100)
}
\description{
L1-penalized SCCA with prespecified sparsity levels for the coefficient vectors
}
\details{
Sparse CCA with L1 penalty is performed repeatedly, while the tuning
parameters are optimized to yield a result that is close to a
prespecified sparsity
}
