% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_latent_variable_data}
\alias{simulate_latent_variable_data}
\title{Simulate data for CCA based on a latent variable model}
\usage{
simulate_latent_variable_data(n_row, n_col_X, n_col_Y, n_signif_X, n_signif_Y,
  noise_var = 1, shared_var_wthn = 0, shared_var_btwn = 1)
}
\description{
Simulate data for CCA based on a latent variable model
}
