% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_MVN_data}
\alias{simulate_MVN_data}
\title{Simulate multivariate normal data}
\usage{
simulate_MVN_data(n_row, n_col_X, n_col_Y, Sigma_chol)
}
\description{
Simulate multivariate normal data
}
